pyproject.toml
[project]
name = "fastapi-web-admin-panel"
version = "0.1.0"
description = "FastAPI Web Admin Panel s autentizací a správou uživatelů"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn==0.23.2",
    "sqlmodel==0.0.8",
    "jinja2==3.1.2",
    "python-multipart==0.0.6",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "alembic==1.12.1",
    "aiofiles==23.2.1",
    "httpx==0.25.1",
    "python-dotenv==1.0.0",
]
